feat: flatiter type fix, errstate divide, view dtype, bool/complex/large int handling
- Fix PyFlatIter to register as proper type (Representable trait)
- Add division errstate checking (FloatingPointError on divide='raise')
- Implement ndarray.view(dtype) with byte-level reinterpretation
- Handle large ints in obj_to_ndarray via __float__ fallback
- Add has_inf() and view_as_dtype() to core NdArray

+++ tests/python/test_numeric.py
@@ -1596,7 +1596,7 @@ def test_full_like_basic():
 
 def test_promote_types():
     r = np.promote_types("int32", "float32")
-    assert_eq(str(r), "float32")
+    assert_eq(str(r), "float64")  # NumPy promotes int32+float32 to float64
 
 def test_promote_types_same():
     r = np.promote_types("float64", "float64")
